# openapi/backend-node.openapi.yaml
openapi: 3.1.0
info:
  title: Nemoclaw Backend Node API
  version: "1.0.0"
  description: |
    Authoritative API contract for one Nemoclaw Backend instance (a "backend
    node"). This is the single source of truth for what a backend node
    exposes today and what is committed for future increments.

    Two tiers:

    - **Tier 1 — OpenAI-compatible surface** (`/v1/*`): interoperability with
      any OpenAI-compatible client. Stable once implemented; no breaking
      changes without explicit architectural approval (see AGENTS.md).
    - **Tier 2 — Native management surface**: Nemoclaw-specific power
      (health, lifecycle, admin, discovery, observability). Stability varies
      by endpoint; see each endpoint's `x-implementation-status` and
      description.

    Every path in this document carries an `x-implementation-status` of one
    of: `implemented`, `partial`, `stub`, `planned`, `unscheduled`. See
    docs/api-contract.md for the legend and the architectural rules this
    contract enforces.
  x-decision-records:
    - "The /admin/model/* namespace is the adopted control-plane contract; it
       supersedes the Nemoclaw system spec's flat /models/load,
       /models/unload, /engines/switch paths (spec §6.5). See commit
       61526a9 and docs/architecture.md."
    - "GET /lifecycle (planned) is read-only observability/status only. It
       performs no transitions. /admin/model/* remains the authoritative
       lifecycle CONTROL surface."
servers:
...
paths:
...
  /v1/chat/completions:
    post:
      tags: [openai-compatible]
      summary: Create a chat completion
      operationId: createChatCompletion
      x-implementation-status: implemented
      x-current-behavior: |
        Both streaming and non-streaming chat completion are implemented
        end to end against the configured engine.

        "stream": true returns Server-Sent Events per OpenAI convention:
        a sequence of "data: {chat.completion.chunk}" events terminated by
        "data: [DONE]". The first event carries {"role": "assistant"};
        subsequent events carry "content" (and, for reasoning models,
        "reasoning") deltas; the final chunk carries finish_reason "stop"
        plus "usage". Only engines advertising streaming support accept
        it - TransformersEngine returns HTTP 400 naming itself, since it
        generates a whole response in one call and has no incremental
        path.

        Reasoning models: whether a model reasons is determined from the
        engine's own capability metadata before any token is emitted, not
        guessed from the token text, because a stream cannot retract what
        it already sent. A model that reasons inline (qwen3:30b) has its
        reasoning prefix buffered until the "</think>" marker proves where
        it ends; the answer after it streams incrementally as usual.

        Model-resolution behavior now differs by active engine
        (docs/ollama-engine-design.md Section 1):

        - TransformersEngine (unchanged, out of scope for that decision):
          the response "model" field echoes the request's "model" (or the
          configured default if omitted) as a label. It does NOT select
          which model actually serves the request — there is exactly one
          loaded model per backend instance, chosen at process startup. A
          client requesting a "model" other than the loaded one still gets
          served by the loaded model, labeled with the requested id.
        - OllamaEngine: a request naming a "model" other than this
          instance's configured/servable model is rejected with HTTP 404
          (see the "404" response below) — never silently served by a
          different model. If "model" is omitted or matches the servable
          model, the response "model" field is the actual served model.
        - Either engine: if the engine's backing runtime is unreachable
          (currently only possible for OllamaEngine, when the Ollama
          daemon is down), the request fails with HTTP 503 (see the "503"
          response below) rather than hanging or raising an unhandled 500.
      description: |
        Implemented for both streaming and non-streaming requests; see
        x-current-behavior for the SSE event shape and reasoning handling.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChatCompletionRequest"
      responses:
        "200":
          description: |
            Chat completion. Content type depends on "stream": a single
            JSON object when false, an SSE event stream when true.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChatCompletionResponse"
            text/event-stream:
              schema:
                $ref: "#/components/schemas/ChatCompletionChunk"
        "400":
          description: |
            An empty "messages" array, or "stream": true against an engine
            that cannot stream (TransformersEngine).
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SimpleErrorResponse"
        "404":
          description: |
            "model" names a model this backend instance cannot serve.
            OllamaEngine only, per docs/ollama-engine-design.md Section 1;
            TransformersEngine never returns this today (see
            x-current-behavior).
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ModelNotFoundResponse"
        "422":
          description: Request body failed schema validation.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HTTPValidationError"
        "503":
          description: |
            The engine's backing runtime is unreachable (e.g. the Ollama
            daemon is down), OR the backend is mid-lifecycle-transition
            (lifecycle_state is not "ready"). Transition-time requests are
            rejected, never queued — see /admin/model/*.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SimpleErrorResponse"
  /v1/models:
    get:
      tags: [openai-compatible]
      summary: List available models
      operationId: listModels
      x-implementation-status: implemented
      description: |
        The authoritative model-listing endpoint. There is NO separate
        native /models endpoint — do not add one; Core and adapters must
        use this endpoint for model discovery.

        Returns every model a caller may select: config/config.yaml's
        model.available catalog, filtered to entries belonging to the
        active backend.engine, joined with what the engine knows about
        each at runtime. This is what a frontend model picker should be
        built on — it can grey out entries that are not "pulled" or do not
        "fit" instead of offering choices that fail on selection.

        The catalog is the gate: only listed models may be loaded or
        switched to (POST /admin/model/*). "loaded" marks which one is
        currently serving, so the listing tracks runtime state.

        Behavior change (2026-08-03, approved): this endpoint previously
        returned exactly one entry, the loaded model. It now enumerates
        the selectable catalog, matching OpenAI's own /v1/models
        semantics. The per-entry OpenAI fields (id, object, created,
        owned_by) are unchanged, so OpenAI-compatible clients still parse
        it; "loaded", "pulled", "size_mib" and "fits" are additive
        Nemoclaw extensions.
      responses:
        "200":
          description: OpenAI-compatible list of selectable models.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ModelListResponse"
        "503":
          description: >-
            The engine's backing runtime is unreachable, so which models
            are usable cannot be determined.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SimpleErrorResponse"
...
  /admin/model/load:
    post:
      tags: [native-management]
      summary: Load a model
      operationId: adminModelLoad
      x-implementation-status: implemented
      description: |
        The /admin/* namespace carries NO stability guarantee.

        Loads model_id when nothing is currently loaded. The target is
        validated twice before any runtime change: first against
        config.yaml's model.available catalog (ModelManager), then by the
        engine against its own reality (for OllamaEngine, that the tag is
        already pulled — it never pulls).

        State rules (docs/model-lifecycle-design.md, Concurrency Model):
        - lifecycle_state "ready" and model_id is the loaded model:
          idempotent success, no engine call.
        - lifecycle_state "ready" and model_id differs: 409, directing the
          caller to /admin/model/switch.
        - lifecycle_state "unloaded" or "degraded": transitions through
          "loading" to "ready", draining any in-flight requests first.

        Only engines advertising runtime lifecycle support accept this.
        TransformersEngine returns 501: its CUDA state is owned in-process
        where cleanup is best-effort, so a supervised worker is required
        (deferred). A missing or malformed body fails FastAPI request
        validation first and returns the standard 422.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ModelLifecycleRequest"
      responses:
        "200":
          description: Model loaded (or already loaded).
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LifecycleResultResponse"
        "404":
          description: model_id is not in config.yaml's model.available catalog.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LifecycleErrorResponse"
        "409":
          description: >-
            Either illegal from the current lifecycle_state (typically
            loading a different model while one is already ready), or the
            engine rejected the target in its pre-flight check
            ("model_unavailable" — e.g. the Ollama tag is configured but
            not pulled). Pre-flight rejections change nothing: the
            previously loaded model keeps serving and lifecycle_state is
            unchanged.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LifecycleErrorResponse"
        "501":
          description: The active engine does not support runtime lifecycle.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LifecycleErrorResponse"
        "503":
          description: >-
            The engine runtime was unreachable, or the load failed after
            the pre-flight check had already passed.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LifecycleErrorResponse"
        "422":
          description: Missing or malformed request body.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HTTPValidationError"
  /admin/model/unload:
    post:
      tags: [native-management]
      summary: Unload the current model
      operationId: adminModelUnload
      x-implementation-status: implemented
      description: |
        The /admin/* namespace carries NO stability guarantee.

        Takes no request body. Drains in-flight requests, then transitions
        through "unloading" to "unloaded". Idempotent when already
        unloaded (no engine call). From "degraded" it transitions straight
        to "unloaded" with a best-effort engine call, since the state
        table has no degraded -> unloading edge.

        For OllamaEngine this asks the daemon to evict the model
        (keep_alive: 0). It is best-effort: the backend neither stops nor
        supervises the daemon process, and eviction is not guaranteed to
        be synchronous.

        While unloaded, POST /v1/chat/completions and POST /generate
        return 503.
      responses:
        "200":
          description: Model unloaded (or already unloaded).
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LifecycleResultResponse"
        "501":
          description: The active engine does not support runtime lifecycle.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LifecycleErrorResponse"
        "503":
          description: The engine runtime was unreachable.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LifecycleErrorResponse"
  /admin/model/switch:
    post:
      tags: [native-management]
      summary: Switch to a different model
      operationId: adminModelSwitch
      x-implementation-status: implemented
      description: |
        The /admin/* namespace carries NO stability guarantee.

        Single transition from one ready model to another. Requires
        lifecycle_state "ready"; from any other state this returns 409
        directing the caller to /admin/model/load. The target is validated
        against config.yaml's model.available catalog before any runtime
        change.

        Requests arriving during "switching" are rejected with 503, never
        queued (docs/model-lifecycle-design.md: queuing hides latency
        spikes and can answer from either model).

        OllamaEngine verifies the target tag is present on the daemon
        BEFORE releasing the current model, so a bad target leaves the old
        model serving and the state stays "ready" — the design doc's
        rollback goal, cheap for a daemon-owned runtime. A failure after
        that point leaves lifecycle_state "degraded".
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ModelLifecycleRequest"
      responses:
        "200":
          description: Switched; loaded_model is now the requested model.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LifecycleResultResponse"
        "404":
          description: model_id is not in config.yaml's model.available catalog.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LifecycleErrorResponse"
        "409":
          description: >-
            Either lifecycle_state was not "ready" (no model to switch
            from), or the engine rejected the target in its pre-flight
            check ("model_unavailable" — e.g. the Ollama tag is configured
            but not pulled). Pre-flight rejections change nothing: the
            previous model keeps serving and lifecycle_state stays
            "ready".
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LifecycleErrorResponse"
        "501":
          description: The active engine does not support runtime lifecycle.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LifecycleErrorResponse"
        "503":
          description: >-
            The engine runtime was unreachable, or the switch failed
            after the pre-flight check had already passed, leaving
            lifecycle_state "degraded".
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LifecycleErrorResponse"
        "422":
          description: Missing or malformed request body.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HTTPValidationError"
...
  /generate:
    post:
      tags: [native-management]
      summary: Legacy text generation (compatibility endpoint)
      operationId: generateLegacy
      x-implementation-status: implemented
      deprecated: true
      description: |
        Implemented legacy compatibility endpoint predating the
        OpenAI-compatible surface. Deprecated: prefer
        POST /v1/chat/completions. Kept only for old local callers per
        docs/architecture.md's Compatibility section.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GenerateRequest"
      responses:
        "200":
          description: Generated text.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GenerateResponse"
        "503":
          description: |
            The engine's backing runtime is unreachable (e.g. the Ollama
            daemon is down), OR the backend is mid-lifecycle-transition
            (lifecycle_state is not "ready"). Transition-time requests are
            rejected, never queued — see /admin/model/*.

            GenerateRequest has no "model" field, so the model-resolution
            404 (see POST /v1/chat/completions) does not apply here.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SimpleErrorResponse"
components:
  schemas:
...
    ChatMessage:
      type: object
      required: [role, content]
      properties:
        role:
          type: string
        content:
          type: string
    ChatCompletionRequest:
      type: object
      required: [messages]
      properties:
        model:
          type: [string, "null"]
          description: >-
            Optional. Behavior differs by active engine — see
            x-current-behavior on POST /v1/chat/completions.
            TransformersEngine: echoed back as a label, does not select
            which model serves the request. OllamaEngine: must be omitted
            or equal to this instance's servable model, or the request is
            rejected with HTTP 404 (model_not_found).
        messages:
          type: array
          items:
            $ref: "#/components/schemas/ChatMessage"
          description: >-
            Must contain at least one message. An empty array is accepted
            by the JSON schema but rejected at runtime with HTTP 400
            ("messages must not be empty").
        max_tokens:
          type: [integer, "null"]
        temperature:
          type: [number, "null"]
        stream:
          type: boolean
          default: false
          description: >-
            Contract: SSE per OpenAI convention when true. Current
            behavior: HTTP 400 when true (see x-current-behavior on the
            operation).
        think:
          type: [boolean, "null"]
          description: >-
            Not part of the OpenAI Chat Completions API - a Nemoclaw
            extension. OllamaEngine-only: forwarded as Ollama's top-level
            `think` request field, controlling whether a reasoning-capable
            model (e.g. Qwen3) emits a <think> trace before its answer.
            null/omitted falls back to the configured model.think_default;
            if that is also null, Ollama's own per-model default applies.
            TransformersEngine ignores this field entirely - no equivalent
            concept there.
    ChatCompletionResponse:
      type: object
      required: [id, object, created, model, choices, usage]
      properties:
        id:
          type: string
          example: "chatcmpl-3f9a1b2c4d5e6f7a8b9c0d1e2f3a4b5c"
        object:
          type: string
          const: chat.completion
        created:
          type: integer
          description: Unix epoch seconds.
        model:
          type: string
        choices:
          type: array
          items:
            type: object
            required: [index, message, finish_reason]
            properties:
              index:
                type: integer
                const: 0
              message:
                type: object
                required: [role, content]
                properties:
                  role:
                    type: string
                    const: assistant
                  content:
                    type: string
                    description: >-
                      The assistant's answer only. For reasoning models any
                      hidden thinking is separated out into "reasoning"
                      below, so an OpenAI-compatible client that renders
                      just this field never shows internal monologue or a
                      stray "</think>" marker.
                  reasoning:
                    type: string
                    description: >-
                      Nemoclaw extension. A reasoning model's hidden
                      thinking, taken from Ollama's "message.thinking" when
                      it supplies one, or split off the raw content at the
                      "</think>" marker when the model's chat template
                      leaks it inline (qwen3:30b MoE does this, and also
                      ignores "think": false entirely). Omitted when the
                      model did no reasoning, which is always the case for
                      non-reasoning models and for TransformersEngine.
              finish_reason:
                type: string
                const: stop
                description: >-
                  Always "stop" today; no other finish reasons are
                  produced (no length/content-filter distinction yet).
        usage:
          $ref: "#/components/schemas/Usage"
    ChatCompletionChunk:
      type: object
      description: >-
        One SSE event's payload when "stream": true. Sent as
        "data: {...}\n\n"; the stream ends with the literal
        "data: [DONE]\n\n" sentinel, which is not JSON.
      required: [id, object, created, model, choices]
      properties:
        id:
          type: string
          description: Stable across every chunk of one completion.
        object:
          type: string
          const: chat.completion.chunk
        created:
          type: integer
        model:
          type: string
        choices:
          type: array
          items:
            type: object
            required: [index, delta, finish_reason]
            properties:
              index:
                type: integer
                const: 0
              delta:
                type: object
                description: >-
                  Incremental fragment. The first chunk carries only
                  "role"; later chunks carry "content" or "reasoning"; the
                  final chunk carries an empty delta.
                properties:
                  role:
                    type: string
                    const: assistant
                  content:
                    type: string
                  reasoning:
                    type: string
                    description: >-
                      Nemoclaw extension, same meaning as in the
                      non-streaming response: hidden thinking, kept out of
                      "content".
              finish_reason:
                type: [string, "null"]
                description: null until the final chunk, then "stop".
        usage:
          $ref: "#/components/schemas/Usage"
    Usage:
      type: object
      required: [prompt_tokens, completion_tokens, total_tokens]
      properties:
        prompt_tokens:
          type: integer
        completion_tokens:
          type: integer
        total_tokens:
          type: integer
    ModelListResponse:
      type: object
      required: [object, data]
      properties:
        object:
          type: string
          const: list
        data:
          type: array
          items:
            $ref: "#/components/schemas/ModelObject"
...
    ModelLifecycleRequest:
      type: object
      required: [model_id]
      properties:
        model_id:
          type: string
        persist:
          type: boolean
          default: false
          description: >-
            When true, a successful transition also rewrites model.id in
            config/config.yaml (via ModelManager.select_model) so a restart
            keeps the new model. Default false is runtime-only: the change
            is lost on restart. Nemoclaw extension, not part of any
            OpenAI spec.
    LifecycleResultResponse:
      type: object
      required:
        [status, lifecycle_state, loaded_model, previous_model,
         elapsed_seconds, persisted]
      properties:
        status:
          type: string
          const: ok
        lifecycle_state:
          $ref: "#/components/schemas/LifecycleState"
        loaded_model:
          type: [string, "null"]
          description: Model id loaded after the transition; null after unload.
        previous_model:
          type: [string, "null"]
          description: Model id loaded before the transition, if any.
        elapsed_seconds:
          type: number
          description: >-
            Wall-clock duration of the transition, including any wait for
            in-flight requests to drain.
        persisted:
          type: boolean
          description: Whether config.yaml was rewritten (see request "persist").
        warning:
          type: string
          description: >-
            Present only when the transition succeeded but something is
            worth flagging - currently that the model looks larger than
            the VRAM the engine runtime can reach, so it may run partly on
            CPU. Advisory: the requirement is estimated from the model's
            on-disk size, so this never fails the operation. Absent when
            there is nothing to report.
    LifecycleErrorResponse:
      type: object
      required: [error, detail, lifecycle_state]
      properties:
        error:
          type: string
          enum:
            - model_not_configured
            - model_unavailable
            - lifecycle_conflict
            - lifecycle_not_supported
            - engine_unavailable
            - lifecycle_failed
        detail:
          type: string
        lifecycle_state:
          $ref: "#/components/schemas/LifecycleState"
          description: State at the time the error was returned.
...
    SimpleErrorResponse:
      type: object
      required: [detail]
      properties:
        detail:
          type: string
      description: FastAPI's default HTTPException body shape.
    ModelNotFoundResponse:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [message, type, param, code]
          properties:
            message:
              type: string
              example: >-
                The model 'llama2:7b' does not exist or is not currently
                loaded by this backend instance.
            type:
              type: string
              const: invalid_request_error
            param:
              type: string
              const: model
            code:
              type: string
              const: model_not_found
      description: >-
        OpenAI-style model_not_found error shape, per
        docs/ollama-engine-design.md Section 1. Deliberately not
        SimpleErrorResponse's {"detail": "..."} shape — matches OpenAI's own
        convention for this error instead.
    HTTPValidationError:
      type: object
      properties:
        detail:
          type: array
          items:
            type: object
            properties:
              loc:
                type: array
                items:
                  oneOf:
                    - type: string
                    - type: integer
              msg:
                type: string
              type:
                type: string
      description: FastAPI's standard 422 validation error body shape.